Made the vague 'concrete asset', that being assets in memory, formally known as resources in the documentation and code structure.
Also moved proper disc assets from the pipeline to the engine. This is because the work around for telling the assets how to turn a general resource into its asset form would have involved magic strings, and just no.

Incidentally, this is probably better for the vibe of the code structure, because the asset system wasn't really part of the pipeline. In fact, it kinda of does the pipeline's job for it (but poorly), so it makes sense to get it out of the pipeline system.